Move createNotes and reviseActionData to testdata/actions.ts
